requirements.txt
appdirs==1.4.4
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
h5py==3.8.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.4.0
protobuf==4.21.12
psutil==5.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
pywavelets==1.4.1
pytz==2022.7.1
PyYAML==6.0
qtorch==0.3.0
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
sentry-sdk==1.14.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.13.9
